+++ private/components/molecules/CountryProfiles/Common.js
@@ -2,17 +2,6 @@ import React from 'react';
 import glamorous from 'glamorous';
 import {red, lightBlack} from 'components/theme/semantic';
 
-export const LightBg = glamorous.div({
-  background: '#e9e7e8',
-  paddingTop: '4em',
-  paddingBottom: '4em',
-});
-export const DarkBg = glamorous.div({
-  background: lightBlack,
-  paddingTop: '2em',
-  paddingBottom: '2em',
-  textAlign: 'center'
-});
 export const Lead = glamorous.p({
   marginTop: '1rem',
   fontSize: '1.8rem',
