+++ private/components/molecules/CountryProfiles/PopulationTab.js
@@ -1,23 +1,33 @@
 import { Container, Header, Grid, Icon, Button } from 'semantic-ui-react';
 import React from 'react';
+import {LightBg} from 'components/atoms/Backgrounds';
 
-const Government = () => (
-  <Container>
+const Population = () => (
+  <LightBg>
     <Grid>
       <Grid.Column width={5}>
         <Header
           textAlign="center"
           as="h3"
         >
-          WHAT IS THE TOTAL REVENUE OF UGANDA?
+          WHAT IS THE POPULATION
         </Header>
         <Header
           textAlign="center"
           as="h1"
           color="red"
         >
-          No data
+          39m
         </Header>
+      </Grid.Column>
+      <Grid.Column width={5}>
+        <Header
+          textAlign="center"
+          as="h3"
+        >
+          WHAT IS THE URBAN VS RURAL SPLIT?
+        </Header>
+
       </Grid.Column>
 
       <Grid.Column width={5}>
@@ -25,28 +35,11 @@ const Government = () => (
           textAlign="center"
           as="h3"
         >
-          HOW MUCH GOVERNMENT REVENUE COMES FROM EXTERNAL GRANTS (AID)?
+          WHAT IS THE AGE PROFILE?
         </Header>
-        <Header
-          textAlign="center"
-          as="h1"
-          color="red"
-        >
-          9.9%
-        </Header>
-      </Grid.Column>
-
-      <Grid.Column width={5}>
-        <Header
-          textAlign="center"
-          as="h3"
-        >
-          HOW IS SPENDING ALLOCATED?
-        </Header>
-
       </Grid.Column>
     </Grid>
-  </Container>
+  </LightBg>
 );
 
-export default Government;
+export default Population;
