+++ private/components/molecules/Front/Common.js
