+++ private/components/molecules/Front/tabs/ForwardLooking.js
