+++ private/components/molecules/Front/tabs/Poverty.js
