+++ private/components/molecules/Front/tabs/Vulnerability.js
