+++ private/components/molecules/SpotLight/tabs/Education.js
