+++ private/components/molecules/SpotLight/tabs/Population.js
