+++ private/components/molecules/SpotLight/tabs/Poverty.js
